Export status indicator code to its own functional component.
